requirements_asr.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
attrs==25.4.0
audioread==3.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
ci_sdr==0.0.2
click==8.3.1
ConfigArgParse==1.7.1
decorator==5.2.1
Distance==0.1.3
editdistance==0.8.1
einops==0.8.2
espnet==202511
espnet-tts-frontend==0.0.3
fast_bss_eval==0.1.3
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.12.0
g2p-en==2.1.0
h5py==3.15.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.11
importlib-metadata==4.13.0
inflect==7.5.0
jaconv==0.5.0
jamo==0.4.1
Jinja2==3.1.6
joblib==1.5.3
kaldiio==2.18.1
lazy_loader==0.4
librosa==0.10.2
lightning==2.6.1
lightning-utilities==0.15.3
llvmlite==0.46.0
MarkupSafe==3.0.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.6.1
nltk==3.9.2
numba==0.64.0
numpy==2.0.0
omegaconf==2.3.0
opt_einsum==3.4.0
packaging==26.0
platformdirs==4.9.2
pooch==1.9.0
propcache==0.4.1
protobuf==6.33.5
pycparser==3.0
pypinyin==0.44.0
pytorch-lightning==2.6.1
pyworld==0.3.5
PyYAML==6.0.3
regex==2026.2.19
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
sentencepiece==0.2.0
soundfile==0.13.1
soxr==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0+cu128
torch-complex==0.4.4
torchaudio==2.8.0+cu128
torchmetrics==1.8.2
tqdm==4.67.3
typeguard==4.5.1
typing_extensions==4.15.0
Unidecode==1.4.0
urllib3==2.6.3
yarl==1.22.0
zipp==3.23.0
espnet_model_zoo==0.1.7
openai-whisper==20250625